Default to GitHub Actions integration ID when app_id is null
On some repositories the current branch protections are returning
null for the app_id, possibly because the checks have not run recently.

For now we can default to the GHA integration ID and fix it
later if there is a mismatch and the check is from another source.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>